Move `StatusReportRow` to its own file
