+++ XenAdmin/XenAdmin.csproj
@@ -794,6 +794,7 @@
     <Compile Include="Dialogs\InstallCertificateDialog.cs">
       <SubType>Form</SubType>
     </Compile>
+    <Compile Include="Wizards\BugToolWizardFiles\StatusReportRows\StatusReportRow.cs" />
     <Compile Include="Wizards\ConversionWizard\ConversionWizard.cs">
       <SubType>Form</SubType>
     </Compile>
